Replace hacky solution with spies from KGB module
The kgb module works by changing the bytecode(!) for a function. By
doing this it means that _all_ existing references to it will call the
new code, meaning we can change the behaviour of a function in place.

Kind of magic? Yes

Possibly going to break in the future? Also yes.